% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r2readthedocs-package.R
\docType{package}
\name{r2readthedocs-package}
\alias{r2readthedocs}
\alias{r2readthedocs-package}
\title{r2readthedocs: Convert R package documentation to a 'readthedocs' website}
\description{
Convert R package documentation to a 'readthedocs' website.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ropenscilabs/r2readthedocs}
  \item Report bugs at \url{https://github.com/ropenscilabs/r2readthedocs/issues}
}

}
\author{
\strong{Maintainer}: Mark Padgham \email{mark.padgham@email.com}

}
\keyword{internal}
