% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg-skeleton.R
\name{rtd_dummy_pkg}
\alias{rtd_dummy_pkg}
\title{Make a dummy package for testing 'r2readthedocs' functions}
\usage{
rtd_dummy_pkg(base_dir = tempdir(), pkg_name = "demo")
}
\arguments{
\item{base_dir}{Base directory in which package is to be constructed (as a
sub-directory)}

\item{pkg_name}{Name of package, which will become a sub-directory of
\code{base_dir}.}
}
\value{
Path to root directory of resultant package
}
\description{
Make a dummy package for testing 'r2readthedocs' functions
}
