% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make-fns.R
\name{rtd_open}
\alias{rtd_open}
\title{Open the main RTD 'html' page in default browser}
\usage{
rtd_open(path = ".")
}
\arguments{
\item{path}{Path to local R package with documentation to be converted}
}
\description{
Open the main RTD 'html' page in default browser
}
