% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readthedocs.R
\name{r2readthedocs}
\alias{r2readthedocs}
\title{Convert package documentation to \code{readthedocs} format}
\usage{
r2readthedocs(path = here::here(), dev = FALSE, open = TRUE)
}
\arguments{
\item{path}{Path to local R package with documentation to be converted}

\item{dev}{If \code{TRUE}, include function documentation of all dependency
packages in site.}

\item{open}{If \code{TRUE}, open the documentation site in default browser.}
}
\value{
TRUE (invisibly) if documentation successfully converted.
}
\description{
Convert package documentation to \code{readthedocs} format
}
