% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make-fns.R
\name{rtd_build}
\alias{rtd_build}
\title{Build the RTD site}
\usage{
rtd_build(path = ".")
}
\arguments{
\item{path}{Path to local R package with documentation to be converted}
}
\description{
Build the RTD site
}
