% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make-fns.R
\name{rtd_clean}
\alias{rtd_clean}
\title{Clean the RTD \verb{_build} directory}
\usage{
rtd_clean(path = ".")
}
\arguments{
\item{path}{Path to local R package with documentation to be converted}
}
\description{
Clean the RTD \verb{_build} directory
}
